requirements.txt
numpy
aiohttp
pybit
python-telegram-bot==20.7
tomli
pytest
